Renamed: ApacheClients to DefaultClients
