+++ algoliasearch-apache/src/test/java/com/algolia/search/analytics/AnalyticsTest.java
@@ -1,11 +1,11 @@
-package com.algolia.search.apache.analytics;
+package com.algolia.search.analytics;
 
-import static com.algolia.search.apache.IntegrationTestExtension.ALGOLIA_API_KEY_1;
-import static com.algolia.search.apache.IntegrationTestExtension.ALGOLIA_APPLICATION_ID_1;
+import static com.algolia.search.IntegrationTestExtension.ALGOLIA_API_KEY_1;
+import static com.algolia.search.IntegrationTestExtension.ALGOLIA_APPLICATION_ID_1;
 
 import com.algolia.search.AnalyticsClient;
-import com.algolia.search.apache.ApacheAnalyticsClient;
-import com.algolia.search.apache.IntegrationTestExtension;
+import com.algolia.search.DefaultAnalyticsClient;
+import com.algolia.search.IntegrationTestExtension;
 import java.io.IOException;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -14,7 +14,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 class AnalyticsTest extends com.algolia.search.integration.analytics.AnalyticsTest {
 
   private static AnalyticsClient analyticsClient =
-      ApacheAnalyticsClient.create(ALGOLIA_APPLICATION_ID_1, ALGOLIA_API_KEY_1);
+      DefaultAnalyticsClient.create(ALGOLIA_APPLICATION_ID_1, ALGOLIA_API_KEY_1);
 
   AnalyticsTest() {
     super(IntegrationTestExtension.searchClient, analyticsClient);
