+++ algoliasearch-apache/src/test/java/com/algolia/search/client/MultipleOperationsTest.java
@@ -1,6 +1,6 @@
-package com.algolia.search.apache.client;
+package com.algolia.search.client;
 
-import com.algolia.search.apache.IntegrationTestExtension;
+import com.algolia.search.IntegrationTestExtension;
 import org.junit.jupiter.api.extension.ExtendWith;
 
 @ExtendWith({IntegrationTestExtension.class})
