+++ algoliasearch-apache/src/test/java/com/algolia/search/client/TimeoutTest.java
@@ -1,11 +1,11 @@
-package com.algolia.search.apache.client;
+package com.algolia.search.client;
 
-import static com.algolia.search.apache.IntegrationTestExtension.ALGOLIA_API_KEY_1;
-import static com.algolia.search.apache.IntegrationTestExtension.ALGOLIA_APPLICATION_ID_1;
+import static com.algolia.search.IntegrationTestExtension.ALGOLIA_API_KEY_1;
+import static com.algolia.search.IntegrationTestExtension.ALGOLIA_APPLICATION_ID_1;
 
+import com.algolia.search.DefaultSearchClient;
 import com.algolia.search.SearchClient;
 import com.algolia.search.SearchConfig;
-import com.algolia.search.apache.ApacheSearchClient;
 
 class TimeoutTest extends com.algolia.search.integration.client.TimeoutTest {
   protected SearchConfig.Builder createBuilder() {
@@ -13,6 +13,6 @@ class TimeoutTest extends com.algolia.search.integration.client.TimeoutTest {
   }
 
   protected SearchClient createClient(SearchConfig config) {
-    return ApacheSearchClient.create(config);
+    return DefaultSearchClient.create(config);
   }
 }
